+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 * Make dashboard table settings popup bigger to accomodate the conditional formatting.
+
+* Issue **#1872** : Partially fixed to show the 1st 1k chars of the single line raw source. Full fix will come in v7.
+
 * Issue **#1868** : Stop `Stream not found with id=nnn` errors during searching.